Move app to typescript.
