+++ react/package.json
@@ -14,6 +14,9 @@
     "shortid": "2.2.8"
   },
   "devDependencies": {
+    "@types/node": "^10.11.4",
+    "@types/prop-types": "^15.5.6",
+    "@types/react": "^16.4.14",
     "babel": "^6.23.0",
     "babel-jest": "^22.4.3",
     "babel-preset-env": "^1.6.1",
@@ -22,7 +25,11 @@
     "jest": "^22.4.3",
     "react": "^16.3.1",
     "react-dom": "^16.3.1",
-    "react-testing-library": "~2.1.1"
+    "react-testing-library": "~2.1.1",
+    "tslint": "^5.11.0",
+    "tslint-config-vtex": "^2.0.0",
+    "tslint-eslint-rules": "^5.4.0",
+    "typescript": "^3.1.1"
   },
   "jest": {
     "verbose": true,
