+++ react/typings/global.d.ts
@@ -1,0 +1,1 @@
+declare var global: any
